[ADD] delete route for users & route's tags

+++ backend/app/modules/users/router.py
@@ -6,7 +6,7 @@ from sqlmodel import Session
 
 router = APIRouter()
 
-@router.post("/register", response_model=Token)
+@router.post("/register", response_model=Token, tags=["OAuth"])
 async def register(user: UserCreate, session: Session = Depends(get_session)) -> Token:
     db_user = get_user_by_email(session, user.email)
     if db_user is not None:
@@ -18,7 +18,7 @@ async def register(user: UserCreate, session: Session = Depends(get_session)) ->
         token_type="bearer"
     )
 
-@router.post("/login", response_model=Token)
+@router.post("/login", response_model=Token, tags=["OAuth"])
 async def login(login: LoginRequest, session: Session = Depends(get_session)) -> Token:
     db_user = get_user_by_email(session=session, email=login.email)
     if db_user is None:
@@ -33,7 +33,7 @@ async def login(login: LoginRequest, session: Session = Depends(get_session)) ->
 
 
 
-@router.get("/me", response_model=UserView)
+@router.get("/me", response_model=UserView, tags=["User"])
 async def read_me(current_user: User = Depends(get_current_user)):
     return UserView(
         id=current_user.id,
@@ -45,11 +45,19 @@ async def read_me(current_user: User = Depends(get_current_user)):
     )
 
 
-@router.patch("/me", response_model=UserView)
+@router.patch("/me", response_model=UserView, tags=["User"])
 async def patch_me(user_update: UserUpdate, current_user: User = Depends(get_current_user), session: Session = Depends(get_session)):
     return update_user(session=session, user=current_user, user_update=user_update)
 
 
-@router.delete("/admin/{user_email}")
+@router.delete("/me", tags=["User"])
+async def delete_user(current_user: User = Depends(get_current_user), session: Session = Depends(get_session)):
+    if current_user.role == "admin":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Admin account can't be deleted themself")
+    return delete_user_by_email(email=current_user.email, user=current_user, session=session)
+
+@router.delete("/admin/{user_email}", tags=["Admin"])
 async def delete_user(user_email: str, current_user: User = Depends(get_current_user), session: Session = Depends(get_session)):
+    if current_user.role != "admin":
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="You don't have the required authorization")
     return delete_user_by_email(email=user_email, user=current_user, session=session)
